Routes rendering css, only remaining todo is header if authenticated
