+++ public/styles.css
@@ -179,8 +179,7 @@ button:active {
 .footer {
   grid-column: 1/7;
   grid-row: 3/3;
-  height: 80px;
   background-color: #d9d9d9;
   text-align: center;
-  padding-top: 40px;
+  padding: 40px 0px 40px 0px;
   margin-top: 80px; }
