+++ views/edit.hbs
@@ -1,6 +1,6 @@
 <div class='content'>
   <div class='formpostion'>
-    <h1>Upload</h1>
+    <h1>Edit</h1>
     <div class='box' >
       <form class='form' action="/{{image_id}}?_method=PUT" method="post">
             <label class="has-float-label">
