+++ views/home.hbs
@@ -2,19 +2,19 @@
     {{#each photos }}
         {{#if featured}}
             <div class='featured'>
-                <a href='/{{ image_id }}'>
+                <a href='/gallery/{{ image_id }}'>
                     <img src="{{ link }}" />    
                 </a>
             </div>
             <br>
         {{/if}}
     {{/each}}
-        <form method="get" action="/new">
+        <form method="get" action="/gallery/new">
             <button class='button' type='submit'>Upload</button>
         </form>
     <div class='gallery'>
         {{#each photos }}
-            <a class='card' href='/{{ image_id }}'>
+            <a class='card' href='/gallery/{{ image_id }}'>
                 <img class='card-image' src="{{ link }}" />
                 <p> 
                     {{ title }}
