+++ views/login.hbs
@@ -9,7 +9,7 @@
             <span>Username</span>
           </label>
           <label class="has-float-label">
-            <input type="text" name="password" placeholder="Enter your pa$$W0rD" required>
+            <input type="password" name="password" placeholder="Enter your pa$$W0rD" required>
             <span>Password</span>
           </label>
           <input type="submit" value="Submit" required>
