+++ views/partials/header.hbs
@@ -17,6 +17,6 @@
         </div>
 
         {{/if}}
-        <a class='button' href="/login">Login</a>
+        <a class='button' href="/login">Login: {{username}}</a>
     </div>
 </div>
