+++ views/upload.hbs
@@ -2,7 +2,7 @@
   <div class='formpostion'>
     <h1>Upload</h1>
     <div class='box' >
-      <form class='form' action="/new" method="post">
+      <form class='form' action="/gallery/new" method="post">
         {{#if addImage}}
           <label class="has-float-label">
             <input type="text" name="title" placeholder="Image title" required>
